Switch from Nginx to Caddy for proxy
This commit replaces Nginx with Caddy for the reverse proxy set-up, offering automatic SSL-certificate issuing. Accompanying changes include the removal of network mode and addition of ports in docker-compose.yml, as well as adjustments to README.md and other relevant files to reflect the transition to Caddy. TCP/UDP features were removed.