Update environment variable and improve documentation
The environment variable `PROXY_HTTP_UPSTREAMS` has been renamed to `PROXY_HTTP_SERVERS` for clarity. Moreover, corresponding changes were made in the README document to enhance the understanding of the docker-compose file. Besides that, an addition was made to improve the HTTPS certification process, which now suggests using a valid email address to enable automatic HTTPS.